[DE-130] static analysis fixes (#10)
* fixed ComparingUnrelatedTypes

* fixed FinalModifierOnCaseClass

* fixed EmptyInterpolatedString

* disabled scapegoat for tests

* disabled HeaderMatchesChecker

* fixed: If block needs braces

* commons scalastyle fixes

* commons scalastyle fixes

* datasource-3.1 scalastyle fixes

* datasource-3.1 scapegoat fixes

* updated dev-README.md

* datasource-3.1 scapegoat fixes

* datasource-2.4 scalastyle fixes

* datasource-2.4 scalastyle fixes

* datasource-2.4 scapegoat fixes

* reverted toIndexedSeq change

* commons scapegoat fixes

+++ integration-tests/src/test/scala/org/apache/spark/sql/arangodb/datasource/BaseSparkTest.scala
@@ -186,7 +186,7 @@ object BaseSparkTest {
       arangoDB.getUser(user)
     } catch {
       case e: ArangoDBException =>
-        if (e.getResponseCode == 404 && e.getErrorNum == 1703)
+        if (e.getResponseCode.toInt == 404 && e.getErrorNum.toInt == 1703)
           arangoDB.createUser(user, password)
         else throw e
     }
